Debug accel/gyro node.

+++ car/nodes/components/accel/component.py
@@ -147,8 +147,8 @@ class AccelerometerGyroscopeSensorComponent(GpioComponent):
                     self.t0 = time()
                     yield (self.dt,
                            self.yaw,
-                           math.ceil(self.ax * 100) / 100,
-                           math.ceil(self.ay * 100) / 100)
+                           float(math.ceil(self.ax * 100) / 100),
+                           float(math.ceil(self.ay * 100) / 100))
 
                 yield None, None, None, None
